Extract Rooms.tsx from Home.tsx
